restructure utils & redux folders
