+++ src/plugins/subscriptions/unsubscribe.js
@@ -1,5 +1,5 @@
-import { omit } from './helpers'
 import { onHandlers } from './handlers'
+import omit from '../../utils/omit'
 
 export const unsubscribe = (modelName: string, matcher: string) => {
   const unsubscribeFrom = (target, formattedMatcher) => {
