+++ src/utils/omit.js
@@ -1,4 +1,5 @@
-export const omit = (prop: string, obj = {}) =>
+// omit key from an object
+export default (prop: string, obj = {}) =>
   Object.keys(obj).reduce((next, key: string) => {
     if (key !== prop) {
       next[key] = obj[key]
