+++ test/store.test.js
@@ -4,14 +4,14 @@ beforeEach(() => {
 
 describe('createStore:', () => {
   it('no params should create store with state `{}`', () => {
-    const { createStore, getStore } = require('../src/utils/store')
+    const { createStore, getStore } = require('../src/redux/store')
     createStore()
 
     expect(getStore().getState()).toEqual({})
   })
 
   it('initialState `null` should create store with state `null`', () => {
-    const { createStore, getStore } = require('../src/utils/store')
+    const { createStore, getStore } = require('../src/redux/store')
     createStore({
       initialState: null
     })
@@ -20,7 +20,7 @@ describe('createStore:', () => {
   })
 
   it('initialState `{ app: "hello, world" }` should create store with state `{ app: "hello, world" }`', () => {
-    const { createStore, getStore } = require('../src/utils/store')
+    const { createStore, getStore } = require('../src/redux/store')
     createStore({
       initialState: { app: 'hello, world' }
     })
@@ -29,7 +29,7 @@ describe('createStore:', () => {
   })
 
   it('extraReducers should create store with extra reducers', () => {
-    const { createStore, getStore } = require('../src/utils/store')
+    const { createStore, getStore } = require('../src/redux/store')
     const extraReducers = { todos: (state = 999) => state }
     createStore({
       extraReducers
@@ -38,7 +38,7 @@ describe('createStore:', () => {
   })
 
   it('rootReducerEnhancer should run a function over combineReducers', () => {
-    const { createStore, getStore } = require('../src/utils/store')
+    const { createStore, getStore } = require('../src/redux/store')
     const rootReducerEnhancer = () => () => 42
     createStore({
       initialState: 2,
